+++ poolparty/src/poolparty/state_ops/__init__.py
@@ -2,7 +2,7 @@
 
 from .repeat import RepeatOp, repeat
 from .stack import StackOp, stack
-from .state_sample import StateSampleOp, state_sample
+from .sample import SampleOp, sample
 from .state_shuffle import StateShuffleOp, state_shuffle
 from .state_slice import StateSliceOp, state_slice
 from .sync import sync
@@ -13,8 +13,8 @@ __all__ = [
     "sync",
     "state_slice",
     "StateSliceOp",
-    "state_sample",
-    "StateSampleOp",
+    "sample",
+    "SampleOp",
     "state_shuffle",
     "StateShuffleOp",
     "repeat",